// src/types/appwrite.d.ts
export type Users = Models.Row & {
    id: string | null;
    email: string | null;
    name: string | null;
    walletAddress: string | null;
    createdAt: string | null;
    updatedAt: string | null;
}
// ...
export type Comments = Models.Row & {
    noteId: string;
    userId: string;
    content: string;
    createdAt: string;
    parentCommentId: string | null;
}